+++ Helpers/ViewRenderer.cs
@@ -1,0 +1,88 @@
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Abstractions;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.AspNetCore.Mvc.Razor;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.AspNetCore.Mvc.ViewEngines;
+using Microsoft.AspNetCore.Mvc.ViewFeatures;
+using Microsoft.AspNetCore.Routing;
+
+namespace Idevs.Helpers;
+
+public interface IViewRenderer
+{
+    Task<string> RenderViewAsync<TModel>(string viewName, TModel model);
+}
+
+public class ViewRenderer : IViewRenderer
+{
+    private IRazorViewEngine ViewEngine { get; }
+    private ITempDataProvider TempDataProvider { get; }
+    private IServiceProvider ServiceProvider { get; }
+
+    public ViewRenderer(IRazorViewEngine viewEngine, ITempDataProvider tempDataProvider, IServiceProvider serviceProvider)
+    {
+        ViewEngine = viewEngine;
+        TempDataProvider = tempDataProvider;
+        ServiceProvider = serviceProvider;
+    }
+
+    public async Task<string> RenderViewAsync<TModel>(string viewName, TModel model)
+    {
+        var actionContext = GetActionContext();
+        var view = FindView(actionContext, viewName);
+
+        await using var output = new StringWriter();
+        var viewContext = new ViewContext(
+            actionContext,
+            view,
+            new ViewDataDictionary<TModel>(
+                metadataProvider: new EmptyModelMetadataProvider(),
+                modelState: new ModelStateDictionary())
+            {
+                Model = model
+            },
+            new TempDataDictionary(
+                actionContext.HttpContext,
+                TempDataProvider),
+            output,
+            new HtmlHelperOptions()
+        );
+
+        await view.RenderAsync(viewContext);
+        return output.ToString();
+    }
+
+    private IView FindView(ActionContext actionContext, string viewName)
+    {
+        var getViewResult = ViewEngine.GetView(null, viewName, false);
+        if (getViewResult.Success)
+        {
+            return getViewResult.View;
+        }
+
+        var findViewResult = ViewEngine.FindView(actionContext, viewName, false);
+        if (findViewResult.Success)
+        {
+            return findViewResult.View;
+        }
+
+        var searchLocations = getViewResult.SearchedLocations.Concat(findViewResult.SearchedLocations);
+        var errorMessage = string.Join(
+            Environment.NewLine,
+            new[] { $"Unable to find view '{viewName}'. The following locations were searched:" }.Concat(
+                searchLocations));
+        throw new InvalidOperationException(errorMessage);
+    }
+
+    private ActionContext GetActionContext()
+    {
+        var httpContext = new DefaultHttpContext
+        {
+            RequestServices = ServiceProvider
+        };
+
+        return new ActionContext(httpContext, new RouteData(), new ActionDescriptor());
+    }
+}