+++ Services/StaticServiceProvider.cs
@@ -1,0 +1,6 @@
+namespace Idevs.Services;
+
+public static class StaticServiceProvider
+{
+    public static IServiceProvider? Provider { get; set; }
+}